refs #268 export logic -> useLogExport
